moving error handling to errs package
